backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.9.0
torchvision==0.24.0
pillow>=12.0.0
